requirements.txt
google-cloud-bigquery
logzero